requirements.txt
streamlit
google-generativeai
langchain
langchain-community
pyPDF2
chromadb
faiss-cpu
langchain_google_genai